+++ package.json
@@ -1,25 +1,6 @@
 {
-  "main": "./scr/index.ts",
+  "main": "./src/index.ts",
   "dependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.12.13",
-    "@babel/plugin-transform-typescript": "^7.12.16",
-    "@babel/preset-env": "^7.12.16",
-    "@types/bcrypt": "^3.0.0",
-    "@types/chalk": "^2.2.0",
-    "@types/connect-redis": "^0.0.16",
-    "@types/cors": "^2.8.9",
-    "@types/dotenv": "^8.2.0",
-    "@types/express-rate-limit": "^5.1.1",
-    "@types/faker": "^5.1.6",
-    "@types/glob": "^7.1.3",
-    "@types/ioredis": "^4.19.3",
-    "@types/node": "^14.14.25",
-    "@types/nodemailer": "^6.4.0",
-    "@types/nodemon": "^1.19.0",
-    "@types/rate-limit-redis": "^1.7.1",
-    "@types/request-promise": "^4.1.47",
-    "@types/uuid": "^8.3.0",
-    "@types/yup": "^0.29.11",
     "babel-jest": "^26.6.3",
     "bcrypt": "^5.0.0",
     "chalk": "^4.1.0",
@@ -49,12 +30,30 @@
     "yup": "^0.32.8"
   },
   "scripts": {
-    "start-server": "NODE_ENV=development nodemon --exec \"ts-node\" scr/index.ts",
+    "start-server": "nodemon --exec \"ts-node\" src/index.ts",
     "session-secret": "ts-node scr/scripts/sessionSecret.ts",
-    "test": "NODE_ENV=test jest --runInBand --detectOpenHandles --colors --verbose ./scr/modules/user/register/test.test.ts"
+    "test": "jest --runInBand --detectOpenHandles --colors --verbose"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.12.13",
+    "@babel/plugin-transform-typescript": "^7.12.16",
+    "@babel/preset-env": "^7.12.16",
+    "@types/bcrypt": "^3.0.0",
+    "@types/chalk": "^2.2.0",
+    "@types/connect-redis": "^0.0.16",
+    "@types/cors": "^2.8.9",
+    "@types/dotenv": "^8.2.0",
+    "@types/express-rate-limit": "^5.1.1",
+    "@types/faker": "^5.1.6",
+    "@types/glob": "^7.1.3",
+    "@types/ioredis": "^4.19.3",
     "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.25",
+    "@types/nodemailer": "^6.4.0",
+    "@types/nodemon": "^1.19.0",
+    "@types/rate-limit-redis": "^1.7.1",
+    "@types/request-promise": "^4.1.47",
+    "@types/uuid": "^8.3.0",
     "jest": "^26.6.3",
     "ts-jest": "^26.5.1",
     "tsconfig-paths": "^3.9.0",
