--- scr/.example.env
@@ -1,52 +1,0 @@
-# test | development | production 
-NODE_ENV=development
-
-HOST=localhost
-
-SERVER_PORT=4000
-CLIENT_PORT=3000
-
-# typeorm
-## development
-TYPEORM_CONNECTION_DEV=postgres
-TYPEORM_HOST_DEV=localhost
-TYPEORM_PORT_DEV=5432
-TYPEORM_USERNAME_DEV=
-TYPEORM_PASSWORD_DEV=
-TYPEORM_DATABASE_DEV=
-## test
-TYPEORM_CONNECTION_TEST=postgres
-TYPEORM_HOST_TEST=localhost
-TYPEORM_PORT_TEST=5432
-TYPEORM_USERNAME_TEST=
-TYPEORM_PASSWORD_TEST=
-TYPEORM_DATABASE_TEST=
-## production
-TYPEORM_CONNECTION_PROD=postgres
-TYPEORM_HOST_PROD=localhost
-TYPEORM_PORT_PROD=5432
-TYPEORM_USERNAME_PROD=
-TYPEORM_PASSWORD_PROD=
-TYPEORM_DATABASE_PROD=
-
-# nodemailer
-EMAIL_HOST=gmail
-EMAIL_USER=
-EMAIL_EMAIL=
-EMAIL_PASSWORD=
-
-# redis
-REDIS_PORT=6379
-REDIS_HOST=localhost
-REDIS_PASSWORD=
-# 1 Day 60 * 60 * 24 - Email token
-REDIS_TOKEN_EXPIRATION=86400
-# 20 Minutes 60 * 20 - Forgot Password Email token
-REDIS_FORGOT_PASSWORD_EXPIRATION=1200
-
-SEED_AMOUNT=3
-SALT_ROUNDS=10
-# generate session secret with: 'yarn session-secret' 
-SESSION_SECRET=
-# 1 Day 60 * 60 * 24 - Acces token
-TOKEN_EXPIRESIN=86400
