+++ src/modules/user/register/sendConfirmEmail.ts
