+++ src/modules/user/shared/User.graphql
