+++ src/utils/genSchema.ts
